+++ src/Annotation/DataTables.php
@@ -3,9 +3,12 @@
 
 namespace Igornast\DataTables\Annotation;
 
+use Igornast\DataTables\ListingBuilder\ListingBuilder;
+
 /**
  * Class DataTables
  * @Annotation
+ * @deprecated Add class name by ListingBuilder
  * @package Igornast\DataTables\Annotation
  */
 class DataTables
@@ -14,9 +17,13 @@ class DataTables
      * @var string
      */
     public $entity;
-
     /**
      * @var string
      */
     public $searchField;
+
+    public function __construct()
+    {
+        trigger_error(sprintf('Annotation class "%s" is deprecated  and will be removed in 2.0, pass name by "%s"', __CLASS__, ListingBuilder::class), E_USER_DEPRECATED);
+    }
 }