+++ src/Listing/Listing.php
@@ -4,6 +4,8 @@
 namespace Igornast\DataTables\Listing;
 
 
+use Igornast\DataTables\Exception\ListingInvalidArgumentException;
+
 class Listing implements ListingInterface
 {
     const DEFAULT_TEMPLATE = '@IgornastDataTables/listing/default_listing.html.twig';
@@ -24,22 +26,40 @@ class Listing implements ListingInterface
      */
     private $options;
     /**
-     * @var string
+     * @var null|string
      */
     private $pathName;
     /**
      * @var string
      */
     private $template;
+    /**
+     * @var null|string
+     */
+    private $encryptedEntity;
+    /**
+     * @var null|string
+     */
+    private $mainSearchField;
 
-    public static function create(string $name, string $pathName, array $columns): self
+    public static function create(string $name, ?string $pathName = null, ?string $encryptedEntity = null): self
     {
+        if(is_null($pathName) && is_null($encryptedEntity)) {
+            throw ListingInvalidArgumentException::NullArgumentsGiven();
+        }
+
+        if(is_null($pathName) === false && is_null($encryptedEntity) === false) {
+            throw ListingInvalidArgumentException::NotNullArgumentsGiven();
+        }
+
         $obj = new self();
         $obj->name = $name;
-        $obj->columns = $columns;
+        $obj->columns = [];
+        $obj->encryptedEntity = $encryptedEntity;
         $obj->filters = [];
         $obj->options = [];
         $obj->pathName = $pathName;
+        $obj->mainSearchField = null;
         $obj->template = self::DEFAULT_TEMPLATE;
 
         return $obj;
@@ -70,17 +90,39 @@ class Listing implements ListingInterface
     }
 
     /**
-     * @return string
-     * @
+     * @return null|string
      */
-    public function getPathName(): string
+    public function getPathName(): ?string
     {
         return $this->pathName;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getMainSearchField(): ?string
+    {
+        return $this->mainSearchField;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getEncryptedEntity(): ?string
+    {
+        return $this->encryptedEntity;
     }
 
     public function addColumn(string $field, string $label): self
     {
         $this->columns = array_merge($this->columns, [$field => ['label' => $label]]);
+
+        return $this;
+    }
+
+    public function addMainSearchField(string $field): self
+    {
+        $this->mainSearchField = $field;
 
         return $this;
     }