+++ src/Tests/ListingBuilderTestCase.php
@@ -15,6 +15,13 @@ class ListingBuilderTestCase extends TestCase
         $this->assertSame('awesome_listing_name', $listingBuilder->getListing()->getName());
     }
 
+    public function testInitializeWithEntityName()
+    {
+        $listingBuilder = new ListingBuilder('awesome_listing_name', null, 'encrypted_entity_name');
+
+        $this->assertSame('encrypted_entity_name', $listingBuilder->getListing()->getEncryptedEntity());
+    }
+
     public function testAddPathName()
     {
         $listingBuilder = new ListingBuilder('awesome_listing_name', 'test_path');
@@ -25,12 +32,39 @@ class ListingBuilderTestCase extends TestCase
     public function testAddOneColumn()
     {
         $listingBuilder = (new ListingBuilder('awesome_listing_name', 'test_path'))
+            ->column('testProperty', 'Property_Name');
+
+        $this->assertCount(1, $listingBuilder->getListing()->getColumns());
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation %s
+     */
+    public function testDeprecatedAddOneColumn()
+    {
+        $listingBuilder = (new ListingBuilder('awesome_listing_name', 'test_path'))
             ->addColumn('testProperty', 'Property_Name');
 
         $this->assertCount(1, $listingBuilder->getListing()->getColumns());
     }
 
     public function testAddMoreColumns()
+    {
+        $listingBuilder = (new ListingBuilder('awesome_listing_name', 'test_path'))
+            ->column('testPropertyOne', 'Property_Name')
+            ->column('testPropertyTwo', 'Property_Name')
+            ->column('testPropertyThree', 'Property_Name')
+            ->column('testPropertyFour', 'Property_Name');
+
+        $this->assertCount(4, $listingBuilder->getListing()->getColumns());
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation %s
+     */
+    public function testDeprecatedAddMoreColumns()
     {
         $listingBuilder = (new ListingBuilder('awesome_listing_name', 'test_path'))
             ->addColumn('testPropertyOne', 'Property_Name')